Windows port: experimental isin: fix MSVC narrowing conversions
- isin_device_operation.cpp, isin_program_factory.cpp: add static_cast<uint32_t>
  for uint64_t logical_volume() values placed into a Shape{} braced initializer
  and a std::vector<uint32_t> compile-time-args list, which MSVC treats as
  narrowing.

+++ ttnn/cpp/ttnn/operations/experimental/isin/device/isin_device_operation.cpp
@@ -37,7 +37,7 @@ void IsInDeviceOperation::validate_on_program_cache_miss(
 IsInDeviceOperation::spec_return_value_t IsInDeviceOperation::compute_output_specs(
     const operation_attributes_t&, const tensor_args_t& tensor_args) {
     return {
-        Shape{tensor_args.elements_tensor.logical_volume()},
+        Shape{static_cast<uint32_t>(tensor_args.elements_tensor.logical_volume())},
         {OUTPUT_TENSOR_DATA_TYPE, {OUTPUT_TENSOR_LAYOUT}, tensor_args.elements_tensor.memory_config()},
     };
 }
